requirements.txt
openai-whisper
torch
transformers
librosa
numpy
pydub
psutil
pyyaml
reportlab
soundfile
PySide6
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
rq
pydantic-settings
